requirements.txt
aiogram>=3.10,<4.0
google-genai>=1.0
PyGithub>=2.0,<3.0
python-dotenv>=1.0
aiohttp>=3.9
